Added test settings
This moves the test settings to a vimeoSettings.json file that is stored
in the /bin folder and therefore not in git. If tests are run and the
file does not exist, it will be generated and an exception is thrown.
This allows contributors to customize the token, album, video, and
userId that is used in integration tests.